Incoming patch for propertySpace ----- (propertySpace)
~+FocusListener for the JTextFields in PropertySpace //now the
TextFields actually do something ^^

~+Constructor got a new parameter (DrawComponent canvas) which allowes
propertySpace to repaint after some value was changed.

+(method) public void mkCircleLabel() creates a diameter label instead
of width and height for "Kreis" objects.

+(feature) you can now name Geometric Objects due to changes in
GeometricObject (the addition of set and get objectname methods) and in
propertyspace on the JTextField textField_Name.

-----
(EditorGUI)

~(method) private void addPropertySpace(GeometricObject nodeInfo)

These method now first checks where the used GeometricObject is and
gives the specific Canvas to the Constructor of PropertySpace 

Signed-off-by: Imkerix <Imkerixus@gmail.com>